==> 02_Pandas/02_Operations.ipynb <==
# 02. 데이터 추가/수정/삭제
---
### DataFrame 행과 열 추가

- 열추가 : DataFrame에 새로운 열을 추가하려면 해당 열 이름과 데이터를 입력한다.

import pandas as pd

data = {
    'Name': ['Alice', 'Bob', 'Charlie', 'David'],
    'Age': [25, 30, 35, 40],
    'City': ['New York', 'Los Angeles', 'Chicago', 'Houston']
}

df = pd.DataFrame(data) 

df

# 'Country' 열 추가

df['Country'] = ['USA', 'USA', 'USA', 'USA']

df

- 행추가 : 새로운 행을 추가하려면 loc를 사용 할 수 있다.

new_row = ['Eve',22, "San Francisco", "USA"]

df.loc[3] = new_row

df

### DataFrame 행과 열 삭제

- 열삭제 : DataFrame에서 열을 삭제하려면 drop() 메서드를 사용한다.
> axis 옵션
> - axis=0 은 행(row)를 의미
> - axis=1 은 열(column)을 의미

df = df.drop('Country', axis=1)

df

- axis를 활용한 행 삭제

df = df.drop(4, axis=0)

df

- 행 삭제 : 특정 행을 삭제하려면 drop() 메서드를 사용하고, 삭제 할 인덱스를 전달한다.

df = df.drop(3)

df

### 기초적인 데이터 수정 및 조작

# df 초기화 
data = {
    'Name': ['Alice', 'Bob', 'Charlie', 'David'],
    'Age': [25, 30, 35, 40],
    'City': ['New York', 'Los Angeles', 'Chicago', 'Houston']
}

df = pd.DataFrame(data)

df

- 데이터 수정 : 특정 값을 변경하려면 인덱스와 열을 지정하고 값을 할당한다.

# 'Age' 열의 값을 수정
df.loc[1,'Age'] = 31

df

- 데이터 타입 변경 : 데이터의 타입을 변경하려면 astype() 메서드를 활용한다.

# 변환하기전 기존 데이터 타입 확인
print(df.dtypes)

df['Age'] = df['Age'].astype(float) # Age의 타입을 floast(실수형)으로 변경

df.dtypes

df
