==> 02_Pandas/03_Index.ipynb <==
# 03. Index
---
- 행(row)에 붙이는 이름표
- 원하는 데이터를 찾을 수 있는 주소

lst = ['유재석','하하']
lst[0]

import pandas as pd

data = {
    '고객ID': ['C001', 'C002', 'C003', 'C004', 'C005', 'C006', 'C007', 'C008'],
    '고객명': ['박태근', '이영희', '박지성', '최민아', '정수빈', '윤태영', '한유진', '강민호'],
    '나이': [35, 30, 35, 40, 28, 27, 31, 29],
    '거주도시': ['시흥', '안양', '울산', '창원', '구로', '시흥', '파주', '일산'],
    '주요관심사': ['Electronics', 'fashion', '', 'Fashion', 'Fashion', 'Electronics', '', 'Sports'],
    '최근1년_방문빈도': [75, 85, 95, 92, 82, 88, 68, 98],
    '평균구매액(만원)': [70, 92, 110, 105, 88, 95, 65, 125],
    '고객만족도(점)': [78, 88, 91, 82, 94, 85, 77, 96],
    '재구매의사(점)': [92, 85, 88, 96, 79, 91, 83, 90]
}

df = pd.DataFrame(data, index=['1번', '2번', '3번', '4번', '5번', '6번', '7번', '8번'])
df

# index를 확인하는 방법
df.index

- Index에 이름을 설정 할 수 있다.

df.index.name = '고객번호'

df

reset_index()
- 인덱스를 초기화할 수 있다.
- 기본 동작은 데이터는 남아있고, 일반 컬럼으로 변한다.

df.reset_index() # index가 일반컬럼으로 바뀜

# 인덱스를 완전히 삭제하려면 drop=True
df.reset_index(drop=True)

df

**inplace=True 옵션**
- 데이터 프레임의 변경을 원본에도 적용시킬 수 있다.

df.reset_index(drop=True, inplace=True)

df

df #고객번호삭제

**set_index()**
- 특정 열을 인덱스로 지정한다.

df.set_index('나이', inplace=True)

df

Index를 기준으로 정렬 할 수 있다.

df.sort_index() #인덱스 기준으로 오름차순 정렬

df.sort_index(ascending=False) # 인덱스 내림차순
