==> 02_Pandas/08_data_select_advanced.ipynb <==
# 8. iloc와 loc
---

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

### .iloc[] (Integer-location)
- 'Integer location'의 약자로, **정수 위치(0부터 시작하는 순서)**를 기반으로 데이터를 선택(조회)하는 기능
- df.iloc[행위치, 열위치] 형태로 사용하며, 이름(label)이 아닌 순서(position)로만 작동한다.

**행(ROW) 전체 조회하기**

df.iloc[0]

df.iloc[0:5:2]

**단일 값(데이터) 조회하기**

df.iloc[0, 1]

df

df.iloc[4, 2]

### .loc[] (Label-location)
- 'Label location'의 약자로, 이름(Label)을 기반으로 데이터를 선택(조회)하는 기능
- 인덱스 이름(예: 'row_A')과 컬럼 이름(예: 'col_X')을 직접 사용

**행(ROW) 전체 조회하기**

df

df.loc['C001':'C005', '고객명':'주요관심사'] # 라벨명으로 조회할때는 끝값도 포함시킴

**단일 값(데이터) 조회하기**
