==> 02_Pandas/11_Preprocessing.ipynb <==
# 11. 데이터 전처리
---
## 결측값 처리
- 데이터 분석에서 결측값은 종종 문제를 발생시킬 수 있으므로 미리 처리해야한다.
- pandas는 결측값을 처리하기 위한 여러 방법을 제공한다.

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

**결측값 제거 `dropna()`**
- 결측치(NaN, Not a Number)를 행또는 열 기준으로 제거해주는 기능
- 원본에는 영향을 끼치지 않고, 새로운 DataFrame이 반환된다.

df.dropna()

**how**
- 행이나 열의 값들이 얼마나 결측치여야 제거할지 결정
- 행기준(axis=0)이 기본값으로 조건에 부합시 그행을 제거함.
    - any : 하나라도 NaN이면 제거 (기본값)
    - all : 전부 NaN이어야 제거

df.dropna(how='all')

import numpy as np
# 모든값이 NaN인 행 추가
df.loc['C009'] = np.nan
df

df.dropna(how='all')

**subset**
- 결측치를 검사할 특정 열을 지정한다. 지정된 열에 결측치가 있을때만 행을 제거한다.''

df.dropna(subset='나이')

df.dropna(subset='주요관심사')

**결측값 대체 `fillna()`**
- 결측치(NaN, Not a Number)를 특정 값이나 방식으로 채워주는 기능
- 원본에는 영향을 끼치지 않고, 새로운 Data Frame이 반환된다.

df.fillna('') # NaN을 빈 문자열값을도 대체

### 중복 데이터 처리
**중복값 확인 `duplicated()`**

data = {
    'Name' : ['Alice', 'Bob', 'Alice', "David"],
    'Age' : [25,30,25,40]
}

df = pd.DataFrame(data)

print(df.duplicated())

**중복값 제거 `drop_duplicates()`**

df = df.drop_duplicates()

df

### 데이터 정렬

**값을 기준으로 정렬 `sort_values()`**
- 특정 열의 값을 기준으로 데이터를 정렬할 수 있다.
- 기본적으로 오름차순으로 정렬되며, 내림차순은 accending=False 옵션을 사용한다.

import pandas as pd
df = pd.read_csv('data/customers3.csv', index_col='고객ID')
df

df.sort_values(by='나이')

**인덱스를 기준으로 정렬 `sort_index()`**
